Extract logic to abstract class to reuse the code and avoid duplication
