+++ psalm-baseline.xml
@@ -5,6 +5,12 @@
       <code>$node</code>
     </InvalidArgument>
   </file>
+  <file src="src/Mutator/Util/AbstractAllSubExprNegation.php">
+    <UndefinedPropertyFetch occurrences="2">
+      <code>$node-&gt;left</code>
+      <code>$node-&gt;right</code>
+    </UndefinedPropertyFetch>
+  </file>
   <file src="src/TestFramework/Coverage/JUnit/JUnitTestFileDataProvider.php">
     <InvalidReturnType occurrences="1">
       <code>iterable&lt;string, string&gt;</code>
