+++ src/Mutator/Boolean/LogicalAndAllSubExprNegation.php
@@ -35,25 +35,18 @@ declare(strict_types=1);
 
 namespace Infection\Mutator\Boolean;
 
-use function in_array;
 use Infection\Mutator\Definition;
-use Infection\Mutator\GetMutatorName;
-use Infection\Mutator\Mutator;
 use Infection\Mutator\MutatorCategory;
-use Infection\PhpParser\Visitor\ParentConnector;
+use Infection\Mutator\Util\AbstractAllSubExprNegation;
 use PhpParser\Node;
 
 /**
  * @internal
  *
- * @implements Mutator<Node\Expr\BinaryOp\BooleanAnd>
+ * @extends AbstractAllSubExprNegation<Node\Expr\BinaryOp\BooleanAnd>
  */
-final class LogicalAndAllSubExprNegation implements Mutator
+final class LogicalAndAllSubExprNegation extends AbstractAllSubExprNegation
 {
-    use GetMutatorName;
-
-    private const BOOLEANS = ['true', 'false'];
-
     public static function getDefinition(): ?Definition
     {
         return new Definition(
@@ -70,89 +63,8 @@ DIFF
         );
     }
 
-    /**
-     * @psalm-mutation-free
-     *
-     * @param Node\Expr|Node\Expr\BinaryOp\BooleanOr $node
-     *
-     * @return iterable<Node>
-     */
-    public function mutate(Node $node): iterable
+    protected function getSupportedBinaryOpExprClass(): string
     {
-        yield $this->negateEverySubExpression($node);
-    }
-
-    public function canMutate(Node $node): bool
-    {
-        if (!$node instanceof Node\Expr\BinaryOp\BooleanAnd) {
-            return false;
-        }
-
-        $parent = ParentConnector::findParent($node);
-
-        // only grandparent
-        if ($parent === null || $parent instanceof Node\Expr\BinaryOp\BooleanAnd) {
-            return false;
-        }
-
-        if ($this->allSubConditionsAreNotMutable($node->left, $node->right)) {
-            return false;
-        }
-
-        return true;
-    }
-
-    private function negateEverySubExpression(Node\Expr|Node\Expr\BinaryOp\BooleanOr $node): Node\Expr
-    {
-        if ($node instanceof Node\Expr\BinaryOp\BooleanAnd) {
-            return new Node\Expr\BinaryOp\BooleanAnd(
-                $this->negateEverySubExpression($node->left),
-                $this->negateEverySubExpression($node->right),
-                $node->getAttributes()
-            );
-        }
-
-        // do not mutate `$a === false` to `!($a === true)` as this is a duplicate of FalseValue mutator
-        if ($this->isIdenticalComparisonWithBoolean($node)) {
-            return $node;
-        }
-
-        return $node instanceof Node\Expr\BooleanNot ? $node->expr : new Node\Expr\BooleanNot($node);
-    }
-
-    /**
-     * `false === $a` or `$a === true` or `$b !== true` etc.
-     */
-    private function isIdenticalComparisonWithBoolean(Node\Expr|Node\Expr\BinaryOp\BooleanOr $node): bool
-    {
-        if (!$node instanceof Node\Expr\BinaryOp\Identical && !$node instanceof Node\Expr\BinaryOp\NotIdentical) {
-            return false;
-        }
-
-        return $this->isBoolean($node->left) || $this->isBoolean($node->right);
-    }
-
-    private function isBoolean(Node\Expr $node): bool
-    {
-        return $node instanceof Node\Expr\ConstFetch && in_array($node->name->toLowerString(), self::BOOLEANS, true);
-    }
-
-    /**
-     * For example if all of them are identical comparisons: `$a === true && $b === false`, we shouldn't mutate
-     */
-    private function allSubConditionsAreNotMutable(Node\Expr $left, Node\Expr $right): bool
-    {
-        $leftIsNotMutable = $left instanceof Node\Expr\BinaryOp\BooleanAnd
-            ? $this->allSubConditionsAreNotMutable($left->left, $left->right)
-            : $this->isIdenticalComparisonWithBoolean($left);
-
-        // optimization: if left is mutable - no need to check right
-        if (!$leftIsNotMutable) {
-            return false;
-        }
-
-        return $right instanceof Node\Expr\BinaryOp\BooleanAnd
-            ? $this->allSubConditionsAreNotMutable($right->left, $right->right)
-            : $this->isIdenticalComparisonWithBoolean($right);
+        return Node\Expr\BinaryOp\BooleanAnd::class;
     }
 }
